+++ tests/Similar_textTest.php
@@ -3,7 +3,11 @@ namespace Ezama\tests{
     require($DIR = (($dir = dirname(__DIR__)).DIRECTORY_SEPARATOR.'src'.DIRECTORY_SEPARATOR)).'similar_text.php';
     require $DIR.'simpleCommonTextSimilarities.php';
     require $DIR.'complexCommonTextSimilarities.php';
-     require $DIR.'distance.php';
+    require $DIR.'distance.php';
+    require $DIR.'diceDistance.php';
+    require $DIR.'jaroWinklerDistance.php';
+    require $DIR.'hammingDistance.php';
+    require $DIR.'levenshteinDistance.php';
     require $dir.DIRECTORY_SEPARATOR.'similar_text.php';
 
     use PHPUnit\Framework\TestCase;
